+++ src/pybamm/models/full_battery_models/lithium_ion/electrode_soh_composite.py
@@ -3,7 +3,7 @@
 #
 import pybamm
 
-from .util import check_if_composite
+from .util import _get_lithiation_delithiation, check_if_composite
 
 
 def _get_stoich_variables(options):
@@ -104,16 +104,20 @@ def _get_cyclable_lithium_equation(options, soc="100"):
 
 
 class ElectrodeSOHComposite(pybamm.BaseModel):
-    """Model to calculate electrode-specific SOH for a cell with composite electrodes, adapted from
-    :footcite:t:`Mohtat2019`.
-    This model is mainly for internal use, to calculate summary variables in a
+    """Model to calculate electrode-specific SOH for a cell with composite electrodes,
+    adapted from :footcite:t:`Mohtat2019`. This model is mainly for internal use, to
+    calculate summary variables in a
     simulation.
 
-    Subscript w indicates working electrode and subscript c indicates counter electrode.
+    Subscript 1 indicates primary phase and subscript 2 indicates secondary phase.
     """
 
     def __init__(
-        self, options, name="ElectrodeSOH model", initialization_method="voltage"
+        self,
+        options,
+        direction=None,
+        name="ElectrodeSOH model",
+        initialization_method="voltage",
     ):
         pybamm.citations.register("Mohtat2019")
         super().__init__(name)
@@ -130,32 +134,98 @@ class ElectrodeSOHComposite(pybamm.BaseModel):
         y_0_1 = variables["y_0_1"]
         V_max = param.voltage_high_cut
         V_min = param.voltage_low_cut
+        # Here we use T_ref as the stoichiometry limits are defined using the reference
+        # state
         if is_negative_composite:
             x_100_2 = variables["x_100_2"]
             x_0_2 = variables["x_0_2"]
-            self.algebraic[x_100_2] = param.n.prim.U(
-                x_100_2, param.T_ref
-            ) - param.n.sec.U(x_100_1, param.T_ref)
-            self.algebraic[x_0_2] = param.n.prim.U(x_0_2, param.T_ref) - param.n.sec.U(
-                x_0_1, param.T_ref
+            self.algebraic[x_100_2] = param.n.sec.U(
+                x_100_2,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="secondary"
+                ),
+            ) - param.n.prim.U(
+                x_100_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="primary"
+                ),
+            )
+            self.algebraic[x_0_2] = param.n.sec.U(
+                x_0_2,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="secondary"
+                ),
+            ) - param.n.prim.U(
+                x_0_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="primary"
+                ),
             )
         if is_positive_composite:
             y_100_2 = variables["y_100_2"]
             y_0_2 = variables["y_0_2"]
-            self.algebraic[y_100_2] = param.p.prim.U(
-                y_100_2, param.T_ref
-            ) - param.p.sec.U(y_100_1, param.T_ref)
-            self.algebraic[y_0_2] = param.p.prim.U(y_0_2, param.T_ref) - param.p.sec.U(
-                y_0_1, param.T_ref
+            self.algebraic[y_100_2] = param.p.sec.U(
+                y_100_2,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="secondary"
+                ),
+            ) - param.p.prim.U(
+                y_100_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="primary"
+                ),
+            )
+            self.algebraic[y_0_2] = param.p.prim.U(
+                y_0_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="primary"
+                ),
+            ) - param.p.sec.U(
+                y_0_2,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="secondary"
+                ),
             )
         self.algebraic[x_100_1] = (
-            param.p.prim.U(y_100_1, param.T_ref)
-            - param.n.prim.U(x_100_1, param.T_ref)
+            param.p.prim.U(
+                y_100_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="primary"
+                ),
+            )
+            - param.n.prim.U(
+                x_100_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="primary"
+                ),
+            )
             - V_max
         )
         self.algebraic[x_0_1] = (
-            param.p.prim.U(y_0_1, param.T_ref)
-            - param.n.prim.U(x_0_1, param.T_ref)
+            param.p.prim.U(
+                y_0_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="primary"
+                ),
+            )
+            - param.n.prim.U(
+                x_0_1,
+                param.T_ref,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="primary"
+                ),
+            )
             - V_min
         )
         # arbitrary choice: use y_0_1 for the capacity equation
@@ -167,10 +237,24 @@ class ElectrodeSOHComposite(pybamm.BaseModel):
         x_init_1 = variables["x_init_1"]
         y_init_1 = variables["y_init_1"]
         if initialization_method == "voltage":
+            # Here we use T_init so that the initial voltage is correct, including the
+            # contribution from the entropic change
             V_init = pybamm.InputParameter("V_init")
             self.algebraic[x_init_1] = (
-                param.p.prim.U(y_init_1, param.T_ref)
-                - param.n.prim.U(x_init_1, param.T_ref)
+                param.p.prim.U(
+                    y_init_1,
+                    param.T_init,
+                    _get_lithiation_delithiation(
+                        direction, "positive", options, phase="primary"
+                    ),
+                )
+                - param.n.prim.U(
+                    x_init_1,
+                    param.T_init,
+                    _get_lithiation_delithiation(
+                        direction, "negative", options, phase="primary"
+                    ),
+                )
                 - V_init
             )
             self.algebraic[y_init_1] = (
@@ -198,17 +282,41 @@ class ElectrodeSOHComposite(pybamm.BaseModel):
             )
         else:
             raise ValueError("Invalid initialization method")
-        # Add voltage equations for secondary phases (init)
+        # Add voltage equations for secondary phases (init), we use T_ref if setting
+        # based on SOC since the stoichiometry limits are defined using the reference
+        # state, and T_init if setting based on voltage since the entropic change is
+        # included in the voltage equation
+        T = param.T_init if initialization_method == "voltage" else param.T_ref
         if is_positive_composite:
             y_init_2 = variables["y_init_2"]
             self.algebraic[y_init_2] = param.p.prim.U(
-                y_init_1, param.T_ref
-            ) - param.p.sec.U(y_init_2, param.T_ref)
+                y_init_1,
+                T,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="primary"
+                ),
+            ) - param.p.sec.U(
+                y_init_2,
+                T,
+                _get_lithiation_delithiation(
+                    direction, "positive", options, phase="secondary"
+                ),
+            )
         if is_negative_composite:
             x_init_2 = variables["x_init_2"]
             self.algebraic[x_init_2] = param.n.prim.U(
-                x_init_1, param.T_ref
-            ) - param.n.sec.U(x_init_2, param.T_ref)
+                x_init_1,
+                T,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="primary"
+                ),
+            ) - param.n.sec.U(
+                x_init_2,
+                T,
+                _get_lithiation_delithiation(
+                    direction, "negative", options, phase="secondary"
+                ),
+            )
 
         self.variables.update(variables)
         if initialization_method == "SOC":
@@ -220,12 +328,13 @@ class ElectrodeSOHComposite(pybamm.BaseModel):
     @property
     def default_solver(self):
         # Use AlgebraicSolver as CasadiAlgebraicSolver gives unnecessary warnings
-        return pybamm.AlgebraicSolver(tol=1e-7)
+        return pybamm.AlgebraicSolver()
 
 
 def get_initial_stoichiometries_composite(
     initial_value,
     parameter_values,
+    direction=None,
     param=None,
     options=None,
     tol=1e-6,
@@ -236,14 +345,25 @@ def get_initial_stoichiometries_composite(
     """
     Get the minimum and maximum stoichiometries from the parameter values
 
-    Parameters
-    ----------
-    parameter_values : pybamm.ParameterValues
-        The parameter values to use in the calculation
-    options : dict, optional
-        A dictionary of options to be passed to the parameters, see
-        :class:`pybamm.BatteryModelOptions`.
-        If None, the default is used: {"working electrode": "positive"}
+        Parameters
+        ----------
+        initial_value : float
+            Target initial value.
+            If integer, interpreted as SOC, must be between 0 and 1.
+            If string e.g. "4 V", interpreted as voltage,
+            must be between V_min and V_max.
+        direction : str, optional
+            The OCV branch to use in the electrode SOH model. Can be "charge" or
+            "discharge".
+        tol : float, optional
+            The tolerance for the solver used to compute the initial stoichiometries.
+            A lower value results in higher precision but may increase computation time.
+            Default is 1e-6.
+        inputs : dict, optional
+            A dictionary of input parameters passed to the model.
+        known_value : str, optional
+            The known value needed to complete the electrode SOH model.
+            Can be "cyclable lithium capacity".
     """
     inputs = inputs or {}
     if known_value != "cyclable lithium capacity":
@@ -277,7 +397,11 @@ def get_initial_stoichiometries_composite(
         all_inputs["SOC_init"] = initial_value
     else:
         raise ValueError("Invalid initial value")
-    model = ElectrodeSOHComposite(options, initialization_method=initialization_method)
-    sim = pybamm.Simulation(model, parameter_values=parameter_values)
+    model = ElectrodeSOHComposite(
+        options, direction, initialization_method=initialization_method
+    )
+    sim = pybamm.Simulation(
+        model, parameter_values=parameter_values, solver=pybamm.AlgebraicSolver(tol=tol)
+    )
     sol = sim.solve([0, 1], inputs=all_inputs)
     return {var: sol[var].entries[0] for var in model.variables.keys()}