+++ src/pybamm/models/full_battery_models/lithium_ion/util.py
@@ -12,3 +12,56 @@ def check_if_composite(options, electrode):
         if electrode == "positive":
             return True
     return False
+
+
+def _has_hysteresis_composite(options, electrode, phase, hysteresis_options):
+    ocp_index = 0 if electrode == "negative" else 1
+    my_ocp_options = options["open-circuit potential"][ocp_index]
+    if check_if_composite(options, electrode) and isinstance(my_ocp_options, tuple):
+        if phase == "primary":
+            if my_ocp_options[0] in hysteresis_options:
+                return True
+            else:
+                return False
+        elif phase == "secondary":
+            if my_ocp_options[1] in hysteresis_options:
+                return True
+            else:
+                return False
+        else:
+            raise ValueError("Phase must be either primary or secondary")
+
+
+def _get_lithiation_delithiation(direction, electrode, options, phase=None):
+    if direction is None or not _has_hysteresis(electrode, options, phase):
+        return None
+    elif (direction == "charge" and electrode == "negative") or (
+        direction == "discharge" and electrode == "positive"
+    ):
+        return "lithiation"
+    elif (direction == "discharge" and electrode == "negative") or (
+        direction == "charge" and electrode == "positive"
+    ):
+        return "delithiation"
+    else:
+        raise ValueError
+
+
+def _has_hysteresis(electrode, options, phase=None):
+    hysteresis_options = [
+        "current sigmoid",
+        "one-state hysteresis",
+        "one-state differential capacity hysteresis",
+    ]
+    phase = phase or ""
+    if options.get("open-circuit potential") is None:
+        return False
+    if isinstance(options["open-circuit potential"], str):
+        if options["open-circuit potential"] in hysteresis_options:
+            return True
+        else:
+            return False
+    elif isinstance(options["open-circuit potential"], tuple):
+        return _has_hysteresis_composite(options, electrode, phase, hysteresis_options)
+    else:
+        raise ValueError